added svg, print options, table styles

+++ src/components/pages/CustomersPage.jsx
@@ -4,6 +4,8 @@ import CustomerModal from "../modal/CustomerModal";
 import toast from "react-hot-toast";
 import { useRouter } from "next/navigation";
 import getActiveOrg from "@/utils/getActiveOrg.mjs";
+import EditSVG from "../svg/EditSVG";
+import DeleteSVG from "../svg/DeleteSVG";
 
 
 const CustomersPage = ({ c, page: p, }) => {
@@ -53,7 +55,7 @@ const CustomersPage = ({ c, page: p, }) => {
     }, [openModal])
     return (
         <div className="w-full">
-            <table className="min-w-full border-collapse border border-gray-300">
+            <table className="item-table">
                 <thead className="bg-gray-200">
                     <tr>
                         <th className="border border-gray-300 p-2 text-left">Name</th>
@@ -74,24 +76,29 @@ const CustomersPage = ({ c, page: p, }) => {
                         >
                             <td className="border border-gray-300 p-2">{c.firstName + " " + c.lastName}
 
-                                <div className="flex space-x-2 opacity-0 group-hover:opacity-100">
+                                <div className="flex gap-10 opacity-0 group-hover:opacity-100 duration-500">
                                     <button
                                         className="text-blue-500 hover:underline"
                                         onClick={(e) => {
                                             e.stopPropagation();
                                             router.push(`/${activeOrg}/customers/new?id=${c._id}`)
                                         }}
+                                        title="Edit this customer"
                                     >
-                                        Edit
+                                        <EditSVG />
                                     </button>
                                     <button
                                         className="text-red-500 hover:underline"
                                         onClick={(e) => {
                                             e.stopPropagation();
-                                            handleDelete(c?._id)
+                                            const confirmed = window.confirm("Sure to delete this customer?")
+                                            if (confirmed) {
+                                                handleDelete(c?._id)
+                                            }
                                         }}
+                                        title="Delete this customer"
                                     >
-                                        Delete
+                                        <DeleteSVG />
                                     </button>
                                 </div>
